Remove time zone from post date
The home page's link to the latest release is 404ing, even though the
post exists. The release post's date is October 30th, but in production,
the post is being listed as October 31st.

This is likely a time zone issue. The post's date in the markdown file
is October 30th 18:40 -07:00, which is October 31st in UTC.

In the future, we should probably refrain from adding time zones.