Add Japanese translation of the Rails Doctrine
This translation is based entirely on the original Japanese version that
takahashim posted on the Japanese tech site Qiita:
https://qiita.com/takahashim/items/85380dff5eec3bd99a6c

I spoke with takahashim about contributing this to the rails/website repository during RubyKaigi 2025.
Since takahashim is currently busy, I've opened this PR on his behalf.
For that reason, I've credited Masayoshi Takahashi as the translator.

There are a few differences from takahashim's original translation:

- The nine themes are not accompanied by the English text. This is to keep things consistent with other language editions.
- I didn't wrap code in `<code>` tags in the text, in line with the current version of "The Rails Doctrine" on rails/website.
  If needed, this formatting can be added in a separate PR to improve readability.

Here are some translator notes. Please feel free to provide any feedback if updates are needed:

- The Japanese title remains "The Rails Doctrine" to match the original translation.
- The translator's name is shown as Masayoshi Takahashi, using the same name format as on GitHub.
  Let me know if you'd prefer it to appear in kanji instead.

Lastly, the change in `_language.scss` is to adjust the height of the language dropdown to accommodate the addition of Japanese.
The order of the dropdown items is based on language codes.

Co-authored-by: Masayoshi Takahashi <takahashimm@gmail.com>